Revert "[#4001/3996] Feature gitlab mr auto update"
