+++ cla-backend-go/users/service.go
@@ -20,7 +20,6 @@ type Service interface {
 	GetUserByLFUserName(lfUserName string) (*models.User, error)
 	GetUserByUserName(userName string, fullMatch bool) (*models.User, error)
 	GetUserByEmail(userEmail string) (*models.User, error)
-	GetUsersByEmail(userEmail string) ([]*models.User, error)
 	GetUserByGitHubID(gitHubID string) (*models.User, error)
 	GetUserByGitHubUsername(gitlabUsername string) (*models.User, error)
 	GetUserByGitlabID(gitHubID int) (*models.User, error)
@@ -165,14 +164,6 @@ func (s service) GetUserByGitLabUsername(gitLabUsername string) (*models.User, e
 		return nil, errors.New("gitLabUsername is empty")
 	}
 	return s.repo.GetUserByGitLabUsername(gitLabUsername)
-}
-
-// GetUsersByEmail fetches the user by email
-func (s service) GetUsersByEmail(userEmail string) ([]*models.User, error) {
-	if userEmail == "" {
-		return nil, errors.New("userEmail is empty")
-	}
-	return s.repo.GetUsersByEmail(userEmail)
 }
 
 // SearchUsers attempts to locate the user by the searchField and searchTerm fields
