+++ cla-backend-go/v2/gitlab-activity/service_test.go
@@ -10,7 +10,6 @@ import (
 	"testing"
 
 	"github.com/communitybridge/easycla/cla-backend-go/gen/v1/models"
-	"github.com/communitybridge/easycla/cla-backend-go/utils"
 	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
 	"github.com/xanzy/go-gitlab"
@@ -95,7 +94,7 @@ func TestIsUserApprovedForSignature(t *testing.T) {
 				expected: true,
 			},
 		}
-		activityService := NewService(nil, nil, nil, nil, nil, nil, nil, nil, nil)
+		activityService := NewService(nil, nil, nil, nil, nil, nil, nil, nil)
 
 		for _, tc := range testCases {
 			t.Run(tc.name, func(tt *testing.T) {
@@ -122,7 +121,7 @@ func TestPrepareMrCommentContent(t *testing.T) {
 		testCases := []struct {
 			name          string
 			signed        []*gitlab.User
-			missing       []*utils.GatedGitlabUser
+			missing       []*gatedGitlabUser
 			expectedMsgs  []string
 			expectedBadge string
 		}{
@@ -140,8 +139,8 @@ func TestPrepareMrCommentContent(t *testing.T) {
 				signed: []*gitlab.User{
 					{ID: 1, Username: "neo"},
 				},
-				missing: []*utils.GatedGitlabUser{
-					{Err: missingID, User: &gitlab.User{ID: 3, Username: "missing"}},
+				missing: []*gatedGitlabUser{
+					{err: missingID, User: &gitlab.User{ID: 3, Username: "missing"}},
 				},
 				expectedMsgs:  []string{signedContains, missingUserContains},
 				expectedBadge: "cla-missing-id.svg",
@@ -151,8 +150,8 @@ func TestPrepareMrCommentContent(t *testing.T) {
 				signed: []*gitlab.User{
 					{ID: 1, Username: "neo"},
 				},
-				missing: []*utils.GatedGitlabUser{
-					{Err: missingCompanyAffiliation, User: &gitlab.User{ID: 4, Username: "affiliationUser"}},
+				missing: []*gatedGitlabUser{
+					{err: missingCompanyAffiliation, User: &gitlab.User{ID: 4, Username: "affiliationUser"}},
 				},
 				expectedMsgs:  []string{signedContains, missingAffiliationContains},
 				expectedBadge: "cla-confirmation-needed.svg",
@@ -162,8 +161,8 @@ func TestPrepareMrCommentContent(t *testing.T) {
 				signed: []*gitlab.User{
 					{ID: 1, Username: "neo"},
 				},
-				missing: []*utils.GatedGitlabUser{
-					{Err: missingCompanyApproval, User: &gitlab.User{ID: 5, Username: "approvalUser"}},
+				missing: []*gatedGitlabUser{
+					{err: missingCompanyApproval, User: &gitlab.User{ID: 5, Username: "approvalUser"}},
 				},
 				expectedMsgs:  []string{signedContains, missingApprovalContains},
 				expectedBadge: "cla-not-signed.svg",
@@ -172,7 +171,7 @@ func TestPrepareMrCommentContent(t *testing.T) {
 
 		for _, tc := range testCases {
 			t.Run(tc.name, func(tt *testing.T) {
-				result := utils.PrepareMrCommentContent(tc.missing, tc.signed, "https://sign.com")
+				result := PrepareMrCommentContent(tc.missing, tc.signed, "https://sign.com")
 				tt.Logf("the result is : %s", result)
 				parts := strings.Split(result, "<li>")
 				assert.Len(tt, parts, len(tc.expectedMsgs)+1)
@@ -192,7 +191,7 @@ func TestPrepareMrCommentContent(t *testing.T) {
 				}
 
 				for i, p := range parts[1:] {
-					expected := fmt.Sprintf(tc.expectedMsgs[i], utils.GetAuthorInfo(allUsers[i]))
+					expected := fmt.Sprintf(tc.expectedMsgs[i], getAuthorInfo(allUsers[i]))
 					assert.Contains(tt, p, expected)
 				}
