+++ cla-backend-go/v2/gitlab_organizations/service.go
@@ -19,6 +19,7 @@ import (
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/communitybridge/easycla/cla-backend-go/company"
+	"github.com/communitybridge/easycla/cla-backend-go/signatures"
 	"github.com/communitybridge/easycla/cla-backend-go/users"
 	"github.com/communitybridge/easycla/cla-backend-go/v2/repositories"
 
@@ -64,13 +65,6 @@ type ServiceInterface interface {
 	RefreshGitLabOrganizationAuth(ctx context.Context, gitLabOrg *common.GitLabOrganization) (*string, error)
 }
 
-// SignatureRepo is a redefine the interface here to avoid circular dependency issues
-type SignatureRepo interface {
-	GetIndividualSignature(ctx context.Context, claGroupID, userID string, approved, signed *bool) (*models.Signature, error)
-	GetCorporateSignature(ctx context.Context, claGroupID, companyID string, approved, signed *bool) (*models.Signature, error)
-	ActivateSignature(ctx context.Context, signatureID string) error
-}
-
 // Service data modelffGetGitLabOrganizationByID
 type Service struct {
 	repo               RepositoryInterface
@@ -79,12 +73,12 @@ type Service struct {
 	gitLabApp          *gitlabApi.App
 	storeRepo          store.Repository
 	userService        users.Service
-	signatureRepo      SignatureRepo
+	signatureRepo      signatures.SignatureRepository
 	companyRepository  company.IRepository
 }
 
 // NewService creates a new gitlab organization service
-func NewService(repo RepositoryInterface, v2GitRepoService repositories.ServiceInterface, claGroupRepository projects_cla_groups.Repository, storeRepo store.Repository, userService users.Service, signaturesRepo SignatureRepo, companyRepository company.IRepository) ServiceInterface {
+func NewService(repo RepositoryInterface, v2GitRepoService repositories.ServiceInterface, claGroupRepository projects_cla_groups.Repository, storeRepo store.Repository, userService users.Service, signaturesRepo signatures.SignatureRepository, companyRepository company.IRepository) ServiceInterface {
 	return &Service{
 		repo:               repo,
 		v2GitRepoService:   v2GitRepoService,
